+++ rand_distr/src/pert.rs
@@ -38,7 +38,7 @@ use rand::Rng;
 ///
 /// [`Triangular`]: crate::Triangular
 #[derive(Clone, Copy, Debug, PartialEq)]
-#[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Pert<F>
 where
     F: Float,